--- src/clocks/counter/aarch64.rs
@@ -1,16 +1,0 @@
-use core::arch::asm;
-
-#[derive(Clone, Debug, Default)]
-pub struct Counter;
-
-impl Counter {
-    pub fn now(&self) -> u64 {
-        let count: u64;
-
-        unsafe {
-            asm!("mrs {}, cntvct_el0", out(reg) count);
-        }
-
-        count
-    }
-}
